Move functions to Misc namespace
